Bootstrap the primal realm using a facade.
